+++ include/mapnik/webp_io.hpp
@@ -47,9 +47,11 @@ int webp_stream_write(const uint8_t* data, size_t data_size, const WebPPicture* 
     return true;
 }
 
-std::string webp_encoding_error(WebPEncodingError error) {
+std::string webp_encoding_error(WebPEncodingError error)
+{
     std::string os;
-    switch (error) {
+    switch (error)
+    {
         case VP8_ENC_ERROR_OUT_OF_MEMORY: os = "memory error allocating objects"; break;
         case VP8_ENC_ERROR_BITSTREAM_OUT_OF_MEMORY: os = "memory error while flushing bits"; break;
         case VP8_ENC_ERROR_NULL_PARAMETER: os = "a pointer parameter is NULL"; break;
@@ -83,19 +85,20 @@ void save_as_webp(T1& file,
 
     // Add additional tuning
     if (method >= 0) config.method = method;
-    #if (WEBP_ENCODER_ABI_VERSION >> 8) >= 1
+#if (WEBP_ENCODER_ABI_VERSION >> 8) >= 1
     config.lossless = !!lossless;
     config.image_hint = static_cast<WebPImageHint>(image_hint);
+#else
+    #ifdef _MSC_VER
+    #pragma NOTE(compiling against webp that does not support lossless flag)
     #else
-        #ifdef _MSC_VER
-        #pragma NOTE(compiling against webp that does not support lossless flag)
-        #else
-        #warning "compiling against webp that does not support lossless flag"
-        #endif
+    #warning "compiling against webp that does not support lossless flag"
     #endif
+#endif
 
     bool valid = WebPValidateConfig(&config);
-    if (!valid) {
+    if (!valid)
+    {
         throw std::runtime_error("Invalid configuration");
     }
 
@@ -106,9 +109,9 @@ void save_as_webp(T1& file,
     }
     pic.width = image.width();
     pic.height = image.height();
-    #if (WEBP_ENCODER_ABI_VERSION >> 8) >= 1
+#if (WEBP_ENCODER_ABI_VERSION >> 8) >= 1
     pic.use_argb = !!lossless;
-    #endif
+#endif
     int ok = 0;
     if (alpha)
     {
@@ -120,11 +123,11 @@ void save_as_webp(T1& file,
     {
         int stride = sizeof(typename T2::pixel_type) * image.width();
         uint8_t const* bytes = reinterpret_cast<uint8_t const*>(image.getBytes());
-        #if (WEBP_ENCODER_ABI_VERSION >> 8) >= 1
+#if (WEBP_ENCODER_ABI_VERSION >> 8) >= 1
         ok = WebPPictureImportRGBX(&pic, bytes, stride);
-        #else
+#else
         ok = WebPPictureImportRGBA(&pic, bytes, stride);
-        #endif
+#endif
     }
 
     if (!ok)