+++ src/font_engine_freetype.cpp
@@ -233,8 +233,8 @@ face_ptr freetype_engine::create_face(std::string const& family_name)
         if (mem_font_itr != memory_fonts_.end()) // memory font
         {
             FT_Error error = FT_New_Memory_Face(library_,
-                                                (FT_Byte const*) mem_font_itr->second.c_str(), //buffer
-                                                mem_font_itr->second.size(), // size
+                                                reinterpret_cast<FT_Byte const*>(mem_font_itr->second.c_str()),
+                                                static_cast<FT_Long>(mem_font_itr->second.size()), // size
                                                 itr->second.first, // face index
                                                 &face);
 
@@ -253,8 +253,8 @@ face_ptr freetype_engine::create_face(std::string const& family_name)
                 = memory_fonts_.insert(std::make_pair(itr->second.second, buffer));
 
             FT_Error error = FT_New_Memory_Face (library_,
-                                                 (FT_Byte const*) result.first->second.c_str(),
-                                                 buffer.size(),
+                                                 reinterpret_cast<FT_Byte const*>(result.first->second.c_str()),
+                                                 static_cast<FT_Long>(buffer.size()),
                                                  itr->second.first,
                                                  &face);
             if (!error) return boost::make_shared<font_face>(face);
@@ -717,7 +717,7 @@ void text_renderer<T>::render_id(mapnik::value_integer feature_id,
                            feature_id,
                            bit->left,
                            height - bit->top,
-                           itr->properties->halo_radius);
+                           static_cast<int>(itr->properties->halo_radius));
         }
     }
 }