+++ tests/visual_tests/test.py
@@ -170,7 +170,10 @@ files = {
     'style-level-compositing-tiled-0,1':{'sizes':[(512,512)],'bbox':merc_z1_bboxes['0,1']},
     'style-level-compositing-tiled-1,1':{'sizes':[(512,512)],'bbox':merc_z1_bboxes['1,1']},
     'marker-path-expression':{},
-    'map-background-image-compositing':{'sizes':[(512,512)]}
+    'map-background-image-compositing':{'sizes':[(512,512)]},
+    'building-symbolizer-opacity':{'sizes':[(512,512)]},
+    'line-pattern-symbolizer-opacity':{'sizes':[(512,512)]},
+    'dst-over-compositing':{'sizes':[(512,512)]},
     }
 
 class Reporting:
@@ -185,7 +188,7 @@ class Reporting:
         self.overwrite_failures = overwrite_failures
         self.errors = [ #(type, actual, expected, diff, message)
          ]
-        
+
     def result_fail(self, actual, expected, diff):
         self.failed += 1
         if self.quiet:
@@ -199,7 +202,7 @@ class Reporting:
             open(expected, 'wb').write(contents)
         else:
             self.errors.append((self.DIFF, actual, expected, diff, None))
-            
+
     def result_pass(self, actual, expected, diff):
         self.passed += 1
         if self.quiet:
@@ -304,7 +307,7 @@ def render(filename,config, width, height, bbox, scale_factor, reporting):
             return m
         reporting.other_error(filename, repr(e))
         return m
-    
+
     for renderer in renderers:
         # TODO - grid renderer does not support scale_factor yet via python
         if renderer['name'] == 'grid' and scale_factor != 1.0: